requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
anyio==3.7.1
    # via httpcore
asgiref==3.7.2
    # via django
beautifulsoup4==4.12.2
    # via
    #   -r requirements.in
    #   django-feed-reader
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.2.0
    # via requests
django==4.2.3
    # via
    #   -r requirements.in
    #   django-feed-reader
    #   djangorestframework
    #   pwned-passwords-django
django-feed-reader==1.0.5
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
exceptiongroup==1.1.2
    # via anyio
feedparser==6.0.10
    # via django-feed-reader
h11==0.14.0
    # via httpcore
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via pwned-passwords-django
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
mysqlclient==2.2.0
    # via -r requirements.in
pwned-passwords-django==2.0
    # via -r requirements.in
pyrfc3339==1.1
    # via django-feed-reader
pytz==2023.3
    # via
    #   djangorestframework
    #   pyrfc3339
requests==2.31.0
    # via
    #   -r requirements.in
    #   django-feed-reader
sgmllib3k==1.0.0
    # via
    #   django-feed-reader
    #   feedparser
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
soupsieve==2.4.1
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
typing-extensions==4.7.1
    # via asgiref
urllib3==2.0.3
    # via requests
